+++ ADMF/functions/Set-AdmfContext.ps1
@@ -1,6 +1,5 @@
-﻿function Set-AdmfContext
-{
-<#
+﻿function Set-AdmfContext {
+	<#
 	.SYNOPSIS
 		Applies a set of configuration contexts.
 	
@@ -71,11 +70,9 @@
 		$EnableException
 	)
 	
-	begin
-	{
+	begin {
 		#region Utility Functions
-		function Set-Context
-		{
+		function Set-Context {
 			[CmdletBinding()]
 			param (
 				$ContextObject,
@@ -97,17 +94,15 @@
 			$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include Server, Credential
 			$stopParam = @{
 				EnableException = $EnableException
-				Cmdlet		    = $Cmdlet
-				Target		    = $ContextObject
-				StepsUpward	    = 1
+				Cmdlet          = $Cmdlet
+				Target          = $ContextObject
+				StepsUpward     = 1
 			}
 			
 			#region PreImport
-			if (Test-Path "$($ContextObject.Path)\preImport.ps1")
-			{
+			if (Test-Path "$($ContextObject.Path)\preImport.ps1") {
 				try { $null = & "$($ContextObject.Path)\preImport.ps1" @parameters }
-				catch
-				{
+				catch {
 					Clear-DMConfiguration
 					Clear-FMConfiguration
 					Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.PreImport' -StringValues $ContextObject.Name -ErrorRecord $_
@@ -118,28 +113,23 @@
 			
 			#region Forest
 			$forestFields = @{
-				'schema' = (Get-Command Register-FMSchema)
+				'schema'    = (Get-Command Register-FMSchema)
 				'sitelinks' = (Get-Command Register-FMSiteLink)
-				'sites'  = (Get-Command Register-FMSite)
-				'subnets' = (Get-Command Register-FMSubnet)
+				'sites'     = (Get-Command Register-FMSite)
+				'subnets'   = (Get-Command Register-FMSubnet)
 			}
 			
-			foreach ($key in $forestFields.Keys)
-			{
+			foreach ($key in $forestFields.Keys) {
 				if (-not (Test-Path "$($ContextObject.Path)\forest\$key")) { continue }
 				
-				foreach ($file in (Get-ChildItem "$($ContextObject.Path)\forest\$key\" -Recurse -Filter "*.json"))
-				{
+				foreach ($file in (Get-ChildItem "$($ContextObject.Path)\forest\$key\" -Recurse -Filter "*.json")) {
 					Write-PSFMessage -Level Debug -String 'Set-AdmfContext.Context.Loading' -StringValues $ContextObject.Name, $key, $file.FullName
-					try
-					{
-						foreach ($dataSet in (Get-Content $file.FullName | ConvertFrom-Json -ErrorAction Stop | Write-Output | ConvertTo-PSFHashtable -Include $($forestFields[$key].Parameters.Keys)))
-						{
+					try {
+						foreach ($dataSet in (Get-Content $file.FullName | ConvertFrom-Json -ErrorAction Stop | Write-Output | ConvertTo-PSFHashtable -Include $($forestFields[$key].Parameters.Keys))) {
 							& $forestFields[$key] @dataSet -ErrorAction Stop
 						}
 					}
-					catch
-					{
+					catch {
 						Clear-DMConfiguration
 						Clear-FMConfiguration
 						Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.ForestConfig' -StringValues $ContextObject.Name, $key, $file.FullName -ErrorRecord $_
@@ -148,22 +138,18 @@
 				}
 			}
 			
-			if (Test-Path "$($ContextObject.Path)\forest\schemaldif")
-			{
+			if (Test-Path "$($ContextObject.Path)\forest\schemaldif") {
 				$filesProcessed = @()
 
 				#region Process Ldif Configuration
-				foreach ($file in (Get-ChildItem "$($ContextObject.Path)\forest\schemaldif\" -Recurse -Filter "*.json"))
-				{
+				foreach ($file in (Get-ChildItem "$($ContextObject.Path)\forest\schemaldif\" -Recurse -Filter "*.json")) {
 					$jsonData = Get-Content $file.FullName | ConvertFrom-Json
-					foreach ($jsonEntry in $jsonData)
-					{
+					foreach ($jsonEntry in $jsonData) {
 						$targetPath = Join-Path "$($ContextObject.Path)\forest\schemaldif" $jsonEntry.Path
 						if ($filesProcessed -contains $targetPath) { continue }
 
 						try { $ldifItem = Get-Item -Path $targetPath -ErrorAction Stop -Force }
-						catch
-						{
+						catch {
 							Clear-DMConfiguration
 							Clear-FMConfiguration
 							Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.ForestConfig' -StringValues $ContextObject.Name, 'Schema (ldif)', $file.FullName -ErrorRecord $_
@@ -171,19 +157,18 @@
 						}
 
 						$ldifParam = @{
-							Path = $ldifItem.FullName
-							Name = $ldifItem.BaseName
+							Path        = $ldifItem.FullName
+							Name        = $ldifItem.BaseName
 							ContextName = $ContextObject.Name
 						}
-						if ($jsonEntry.Name) {  $ldifParam.Name = $jsonEntry.Name }
+						if ($jsonEntry.Name) { $ldifParam.Name = $jsonEntry.Name }
 						if ($jsonEntry.Weight) { $ldifParam['Weight'] = $jsonEntry.Weight }
 						if ($jsonEntry.MissingObjectExemption) { $ldifParam['MissingObjectExemption'] = $jsonEntry.MissingObjectExemption }
 						try {
 							Register-FMSchemaLdif @ldifParam -ErrorAction Stop
 							$filesProcessed += $ldifItem.FullName
 						}
-						catch
-						{
+						catch {
 							Clear-DMConfiguration
 							Clear-FMConfiguration
 							Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.ForestConfig' -StringValues $ContextObject.Name, 'schemaldif', $file.FullName -ErrorRecord $_
@@ -194,14 +179,12 @@
 				#endregion Process Ldif Configuration
 
 				#region Process Ldif Files without configuration
-				foreach ($file in (Get-ChildItem "$($ContextObject.Path)\forest\schemaldif\" -Recurse -Filter "*.ldf"))
-				{
+				foreach ($file in (Get-ChildItem "$($ContextObject.Path)\forest\schemaldif\" -Recurse -Filter "*.ldf")) {
 					# Skip files already defined in json
 					if ($filesProcessed -contains $file.FullName) { continue }
 					
 					try { Register-FMSchemaLdif -Name $file.BaseName -Path $file.FullName -ContextName $ContextObject.Name -ErrorAction Stop }
-					catch
-					{
+					catch {
 						Clear-DMConfiguration
 						Clear-FMConfiguration
 						Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.ForestConfig' -StringValues $ContextObject.Name, 'schemaldif', $file.FullName -ErrorRecord $_
@@ -214,40 +197,36 @@
 			
 			#region Domain
 			$domainFields = @{
-				'accessrules' = (Get-Command Register-DMAccessRule)
-				'acls'		  = (Get-Command Register-DMAcl)
-				'builtinsids' = (Get-Command Register-DMBuiltInSID)
-				'gplinks'	  = (Get-Command Register-DMGPLink)
-				'gppermissions' = (Get-Command Register-DMGPPermission)
+				'accessrules'         = (Get-Command Register-DMAccessRule)
+				'acls'                = (Get-Command Register-DMAcl)
+				'builtinsids'         = (Get-Command Register-DMBuiltInSID)
+				'gplinks'             = (Get-Command Register-DMGPLink)
+				'gppermissions'       = (Get-Command Register-DMGPPermission)
 				'gppermissionfilters' = (Get-Command Register-DMGPPermissionFilter)
-				'groups'	  = (Get-Command Register-DMGroup)
-				'groupmemberships' = (Get-Command Register-DMGroupMembership)
-				'names'	      = (Get-Command Register-DMNameMapping)
-				'objects'	  = (Get-Command Register-DMObject)
+				'gpregistrysettings'  = (Get-Command Register-DMGPRegistrySetting)
+				'groups'              = (Get-Command Register-DMGroup)
+				'groupmemberships'    = (Get-Command Register-DMGroupMembership)
+				'names'               = (Get-Command Register-DMNameMapping)
+				'objects'             = (Get-Command Register-DMObject)
 				'organizationalunits' = (Get-Command Register-DMOrganizationalUnit)
-				'psos'	      = (Get-Command Register-DMPasswordPolicy)
-				'users'	      = (Get-Command Register-DMUser)
+				'psos'                = (Get-Command Register-DMPasswordPolicy)
+				'users'               = (Get-Command Register-DMUser)
 			}
 			
-			foreach ($key in $domainFields.Keys)
-			{
+			foreach ($key in $domainFields.Keys) {
 				if (-not (Test-Path "$($ContextObject.Path)\domain\$key")) { continue }
 				
-				foreach ($file in (Get-ChildItem "$($ContextObject.Path)\domain\$key\" -Recurse -Filter "*.json"))
-				{
+				foreach ($file in (Get-ChildItem "$($ContextObject.Path)\domain\$key\" -Recurse -Filter "*.json")) {
 					Write-PSFMessage -Level Debug -String 'Set-AdmfContext.Context.Loading' -StringValues $ContextObject.Name, $key, $file.FullName
-					try
-					{
-						foreach ($dataSet in (Get-Content $file.FullName | ConvertFrom-Json -ErrorAction Stop | Write-Output | ConvertTo-PSFHashtable -Include $($domainFields[$key].Parameters.Keys)))
-						{
+					try {
+						foreach ($dataSet in (Get-Content $file.FullName | ConvertFrom-Json -ErrorAction Stop | Write-Output | ConvertTo-PSFHashtable -Include $($domainFields[$key].Parameters.Keys))) {
 							if ($domainFields[$key].Parameters.Keys -contains 'ContextName') {
 								$dataSet['ContextName'] = $ContextObject.Name
 							}
 							& $domainFields[$key] @dataSet -ErrorAction Stop
 						}
 					}
-					catch
-					{
+					catch {
 						Clear-DMConfiguration
 						Clear-FMConfiguration
 						Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.DomainConfig' -StringValues $ContextObject.Name, $key, $file.FullName -ErrorRecord $_
@@ -257,13 +236,12 @@
 			}
 
 			# Group Policy
-			if (Test-Path "$($ContextObject.Path)\domain\grouppolicies\exportData.json")
-			{
+			if (Test-Path "$($ContextObject.Path)\domain\grouppolicies\exportData.json") {
 				$file = Get-Item "$($ContextObject.Path)\domain\grouppolicies\exportData.json"
 				Write-PSFMessage -Level Debug -String 'Set-AdmfContext.Context.Loading' -StringValues $ContextObject.Name, 'Group Policy', $file.FullName
 				try {
 					$dataSet = Get-Content $file.FullName | ConvertFrom-Json -ErrorAction Stop | ConvertTo-PSFHashtable -Include DisplayName, Description, ID, ExportID
-					foreach ($policyEntry in $dataSet){
+					foreach ($policyEntry in $dataSet) {
 						Register-DMGroupPolicy @policyEntry -Path "$($ContextObject.Path)\domain\grouppolicies\$($policyEntry.ID)"
 					}
 				}
@@ -276,9 +254,10 @@
 			}
 
 			# Object Categories
-			foreach ($file in (Get-ChildItem "$($ContextObject.Path)\domain\objectcategories" -Filter '*.psd1')){
+			foreach ($file in (Get-ChildItem "$($ContextObject.Path)\domain\objectcategories" -Filter '*.psd1' -ErrorAction Ignore)) {
 				try {
 					$dataSet = Import-PSFPowerShellDataFile -Path $file.FullName
+					$dataSet.TestScript = $dataSet.TestScript.Invoke() | Write-Output # Remove automatic scriptblock nesting
 					Register-DMObjectCategory @dataSet
 				}
 				catch {
@@ -288,37 +267,46 @@
 					return
 				}
 			}
+
+			# Domain Data
+			foreach ($file in (Get-ChildItem "$($ContextObject.Path)\domain\domaindata" -Filter '*.psd1' -ErrorAction Ignore)) {
+				try {
+					$dataSet = Import-PSFPowerShellDataFile -Path $file.FullName
+					$dataSet.Scriptblock = $dataSet.Scriptblock.Invoke() | Write-Output # Remove automatic scriptblock nesting
+					Register-DMDomainData @dataSet
+				}
+				catch {
+					Clear-DMConfiguration
+					Clear-FMConfiguration
+					Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.DomainConfig' -StringValues $ContextObject.Name, 'Domain Data', $file.FullName -ErrorRecord $_
+					return
+				}
+			}
 			
 			# Content Mode
-			if (Test-Path "$($ContextObject.Path)\domain\content_mode.json")
-			{
+			if (Test-Path "$($ContextObject.Path)\domain\content_mode.json") {
 				$file = Get-Item "$($ContextObject.Path)\domain\content_mode.json"
 				Write-PSFMessage -Level Debug -String 'Set-AdmfContext.Context.Loading' -StringValues $ContextObject.Name, 'ContentMode', $file.FullName
-				try
-				{
+				try {
 					$dataSet = Get-Content $file.FullName | ConvertFrom-Json -ErrorAction Stop
 					if ($dataSet.Mode) { Set-DMContentMode -Mode $dataSet.Mode }
-					if ($dataSet.Include)
-					{
+					if ($dataSet.Include) {
 						$includes = @((Get-DMContentMode).Include)
 						foreach ($entry in $dataSet.Include) { $includes += $entry }
 						Set-DMContentMode -Include $includes
 					}
-					if ($dataSet.Exclude)
-					{
+					if ($dataSet.Exclude) {
 						$excludes = @((Get-DMContentMode).Exclude)
 						foreach ($entry in $dataSet.Exclude) { $excludes += $entry }
 						Set-DMContentMode -Exclude $excludes
 					}
-					if ($dataSet.UserExcludePattern)
-					{
+					if ($dataSet.UserExcludePattern) {
 						$userExcludePatterns = @((Get-DMContentMode).UserExcludePattern)
 						foreach ($entry in $dataSet.UserExcludePattern) { $userExcludePatterns += $entry }
 						Set-DMContentMode -UserExcludePattern $userExcludePatterns
 					}
 				}
-				catch
-				{
+				catch {
 					Clear-DMConfiguration
 					Clear-FMConfiguration
 					Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.DomainConfig' -StringValues $ContextObject.Name, 'ContentMode', $file.FullName -ErrorRecord $_
@@ -328,11 +316,9 @@
 			#endregion Domain
 			
 			#region DC
-			if (Test-Path "$($ContextObject.Path)\dc\dc_config.json")
-			{
+			if (Test-Path "$($ContextObject.Path)\dc\dc_config.json") {
 				try { $dcData = Get-Content "$($ContextObject.Path)\dc\dc_config.json" -ErrorAction Stop | ConvertFrom-Json -ErrorAction Stop }
-				catch
-				{
+				catch {
 					Clear-DMConfiguration
 					Clear-FMConfiguration
 					Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.DCConfig' -StringValues $ContextObject.Name -ErrorRecord $_
@@ -348,11 +334,9 @@
 			#endregion DC
 			
 			#region PostImport
-			if (Test-Path "$($ContextObject.Path)\postImport.ps1")
-			{
+			if (Test-Path "$($ContextObject.Path)\postImport.ps1") {
 				try { $null = & "$($ContextObject.Path)\postImport.ps1" @parameters }
-				catch
-				{
+				catch {
 					Clear-DMConfiguration
 					Clear-FMConfiguration
 					Stop-PSFFunction @stopParam -String 'Set-AdmfContext.Context.Error.PostImport' -StringValues $ContextObject.Name -ErrorRecord $_
@@ -369,24 +353,20 @@
 		# Common parameters for Stop-PSFFunction
 		$commonParam = @{
 			EnableException = $EnableException
-			Continue	    = $true
+			Continue        = $true
 		}
 	}
-	process
-	{
+	process {
 		#region Explicitly specified contexts
-		foreach ($contextObject in $Context)
-		{
-			if ($contextObject -is [string])
-			{
+		foreach ($contextObject in $Context) {
+			if ($contextObject -is [string]) {
 				$foundContext = Get-AdmfContext -Name $contextObject
 				if (-not $foundContext) { Stop-PSFFunction @commonParam -String 'Set-AdmfContext.Context.NotFound' -StringValues $contextObject }
 				if ($foundContext.Count -gt 1) { Stop-PSFFunction @commonParam -String 'Set-AdmfContext.Context.Ambiguous' -StringValues $contextObject, ($foundContext.Name -join ", ") }
 				$selectedContexts[$foundContext.Name] = $foundContext
 				continue
 			}
-			if ($contextObject.PSObject.Typenames -eq 'ADMF.Context')
-			{
+			if ($contextObject.PSObject.Typenames -eq 'ADMF.Context') {
 				$selectedContexts[$contextObject.Name] = $contextObject
 				continue
 			}
@@ -394,45 +374,36 @@
 		}
 		#endregion Explicitly specified contexts
 		#region Interactively chosen contexts
-		if ($Interactive)
-		{
-			if ($ReUse -and $script:assignedContexts["$Server"])
-			{
-				foreach ($contextObject in $script:assignedContexts["$Server"])
-				{
+		if ($Interactive) {
+			if ($ReUse -and $script:assignedContexts["$Server"]) {
+				foreach ($contextObject in $script:assignedContexts["$Server"]) {
 					$selectedContexts[$contextObject.Name] = $contextObject
 				}
 				return
 			}
-			try
-			{
-				foreach ($contextObject in (Invoke-CallbackMenu @parameters))
-				{
+			try {
+				foreach ($contextObject in (Invoke-CallbackMenu @parameters)) {
 					$selectedContexts[$contextObject.Name] = $contextObject
 				}
 			}
-			catch
-			{
+			catch {
 				Stop-PSFFunction -String 'Set-AdmfContext.Interactive.Cancel' -EnableException $EnableException -ErrorRecord $_
 				return
 			}
 		}
 		#endregion Interactively chosen contexts
 	}
-	end
-	{
+	end {
 		if (Test-PSFFunctionInterrupt) { return }
 		
 		#region Handle errors in selection
 		$missingPrerequisites = $selectedContexts.Values.Prerequisites | Where-Object { $_ -notin $selectedContexts.Values.Name }
-		if ($missingPrerequisites)
-		{
+		if ($missingPrerequisites) {
 			Stop-PSFFunction -String 'Set-AdmfContext.Resolution.MissingPrerequisites' -StringValues ($missingPrerequisites -join ", ") -EnableException $EnableException -Category InvalidData
 			return
 		}
 		$conflictingContexts = $selectedContexts.Values.MutuallyExclusive | Where-Object { $_ -in $selectedContexts.Values.Name }
-		if ($conflictingContexts)
-		{
+		if ($conflictingContexts) {
 			Stop-PSFFunction -String 'Set-AdmfContext.Resolution.ExclusionConflict' -StringValues ($conflictingContexts.Name -join ", ") -EnableException $EnableException -Category InvalidData
 			return
 		}
@@ -443,8 +414,7 @@
 			$script:loadedContexts.Name -and
 			$selectedContexts.Values.Name -and
 			-not (Compare-Object -ReferenceObject $selectedContexts.Values.Name -DifferenceObject $script:loadedContexts.Name)
-		)
-		{
+		) {
 			# When switching from one domain to a new one, make sure that the selection is cached, even if it is the same selection.
 			# Otherwise, the second domain will keep reprompting for contexts
 			if (-not $script:assignedContexts["$Server"]) { $script:assignedContexts["$Server"] = $selectedContexts.Values }
@@ -458,8 +428,7 @@
 		
 		Set-PSFTaskEngineCache -Module ADMF -Name currentlyImportingContexts -Value $selectedContexts.Values
 		
-		foreach ($contextObject in ($selectedContexts.Values | Sort-Object Weight))
-		{
+		foreach ($contextObject in ($selectedContexts.Values | Sort-Object Weight)) {
 			if (Test-PSFFunctionInterrupt) { return }
 			Set-Context @parameters -ContextObject $contextObject -Cmdlet $PSCmdlet -EnableException $EnableException
 			if (Test-PSFFunctionInterrupt) { return }